+++ samples/hello/hello_text.rb
@@ -36,8 +36,36 @@ class HelloText
     @background = BACKGROUND_PROMPT
     @font_family = FONT_FAMILY_PROMPT
     @font_size = FONT_SIZE_PROMPT
+    @document = <<~MULTI_LINE_STRING
+      According to the National Post, a heavy metal-loving high school principal in Canada will be allowed to keep her job, days after a public campaign to oust her made headlines around the globe.
+      
+      Parents at Eden High School in St. Catharines, Ontario launched a petition to remove principal Sharon Burns after discovering she's an IRON MAIDEN fan.
+      
+      The petition, titled "Eden High School Principal, Sharon Burns, Needs to Be Transferred Immediately!" read, "We are deeply disturbed that the principal assigned to the school blatantly showed Satanic symbols and her allegiance to Satanic practices on her public social media platforms where all the students can see them under @edenprincipal (not her personal account)."
+      
+      More than 500 people signed the petition to transfer Burns after she posted a picture of herself flashing the "devil horns" hand sign with a doll of the MAIDEN zombie mascot Eddie behind her as well as a personalized license plate reading "IRNMADEN" and a handwritten note that reads "Eddie 666" on a car's dashboard.
+      
+      
+      The number 666 is used to represent the devil, and is featured prominently in MAIDEN's artwork by the band, whose classic third album is titled "The Number Of The Beast".
+      
+      The petition was later updated to state that the demand for her transfer is not because of her love for metal, but for "openly displaying her own handmade sign with the 666 clearly displayed on it".
+      
+      The creator of the original petition wrote: "Sharon knows full well what she did was simply inappropriate, unnecessary and not professional but has yet to publicly admit so and is willing to allow people to believe a completely different story, making very real concerns seem petty."
+      
+      Meanwhile, a counter-petition supporting Burns garnered over 20,000 signatures.
+      
+      "It is ridiculous that a couple of parents judge her role as a principal only based on an Instagram post. (About liking the band IRON MAIDEN. That's it.) Eden High School is a public school. Not a Christian school," the petition titled "We need Mrs Burns" stated. "She has made Eden a safe space for so many people. She spreads nothing but love and kindness."
+      
+      After the complaints were aired, the District School Board of Niagara spoke with Burns and the parents who launched the petition, and the issue is over as far as the board is concerned, Kim Sweeney, the board's chief communications officer, told the National Post. No disciplinary action or policy changes were needed.
+      
+      "Our belief is that taste in music is subjective and we support that both students and staff enjoy a wide variety of genres," Sweeney said.
+      
+      The original petition has since been removed.
+    MULTI_LINE_STRING
   end
   
+  attr_accessor :document
+      
   attr_accessor :foreground
   
   def foreground_options
@@ -201,9 +229,9 @@ class HelloText
             # TkTextImage.new(text, 'sel.first', :image => flowers)
             # Replace everything below with @text.open_file_to_insert_image(text_index = 'insert')
             file_types = [
-              '{PNG Files} {.png}',
-              '{Gif Files} {.gif}',
-              '{PPM Files} {.ppm}',
+              '{PNG Images} {.png}',
+              '{Gif Images} {.gif}',
+              '{PPM Images} {.ppm}',
             ]
             image_filename = ::Tk::getOpenFile(filetypes: file_types)
             image = TkPhotoImage.new(:file => image_filename)
@@ -216,32 +244,7 @@ class HelloText
         grid row: 1, column: 0, row_weight: 1
         wrap 'word'
         undo true
-        value <<~MULTI_LINE_STRING
-          According to the National Post, a heavy metal-loving high school principal in Canada will be allowed to keep her job, days after a public campaign to oust her made headlines around the globe.
-          
-          Parents at Eden High School in St. Catharines, Ontario launched a petition to remove principal Sharon Burns after discovering she's an IRON MAIDEN fan.
-          
-          The petition, titled "Eden High School Principal, Sharon Burns, Needs to Be Transferred Immediately!" read, "We are deeply disturbed that the principal assigned to the school blatantly showed Satanic symbols and her allegiance to Satanic practices on her public social media platforms where all the students can see them under @edenprincipal (not her personal account)."
-          
-          More than 500 people signed the petition to transfer Burns after she posted a picture of herself flashing the "devil horns" hand sign with a doll of the MAIDEN zombie mascot Eddie behind her as well as a personalized license plate reading "IRNMADEN" and a handwritten note that reads "Eddie 666" on a car's dashboard.
-          
-          
-          The number 666 is used to represent the devil, and is featured prominently in MAIDEN's artwork by the band, whose classic third album is titled "The Number Of The Beast".
-          
-          The petition was later updated to state that the demand for her transfer is not because of her love for metal, but for "openly displaying her own handmade sign with the 666 clearly displayed on it".
-          
-          The creator of the original petition wrote: "Sharon knows full well what she did was simply inappropriate, unnecessary and not professional but has yet to publicly admit so and is willing to allow people to believe a completely different story, making very real concerns seem petty."
-          
-          Meanwhile, a counter-petition supporting Burns garnered over 20,000 signatures.
-          
-          "It is ridiculous that a couple of parents judge her role as a principal only based on an Instagram post. (About liking the band IRON MAIDEN. That's it.) Eden High School is a public school. Not a Christian school," the petition titled "We need Mrs Burns" stated. "She has made Eden a safe space for so many people. She spreads nothing but love and kindness."
-          
-          After the complaints were aired, the District School Board of Niagara spoke with Burns and the parents who launched the petition, and the issue is over as far as the board is concerned, Kim Sweeney, the board's chief communications officer, told the National Post. No disciplinary action or policy changes were needed.
-          
-          "Our belief is that taste in music is subjective and we support that both students and staff enjoy a wide variety of genres," Sweeney said.
-          
-          The original petition has since been removed.
-        MULTI_LINE_STRING
+        value <=> [self, :document]
       }
     }.open
   end