+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## 0.0.29
 
+- Improve Hello, Text! (toolbar buttons: insert image)
 - Removed unnecessary `text` widget `text` attribute after realizing `value` does the same thing
 - Support `text` widget data-binding on 'modified' event
 - Support `text#insert_image(text_index, *image_args)`
+- Support built-in dialogs: `get_open_file(options = nil)`, `get_multiple_open_file(options = nil)`, `get_save_file(options = nil)`, `get_multiple_save_file(options = nil)`, `choose_directory(options = nil)`, `choose_color(options = nil)`, `choose_font(initial_font_args) {|font| do_something_with_font(font) }`
+
 
 ## 0.0.28